requirements.txt
Django==5.2.7
mysqlclient==2.2.7
python-dotenv==1.1.1
Pillow
weasyprint
qrcode
django-jazzmin==3.0.1
djangorestframework==3.15.1